requirements.txt
Django==4.2
djangorestframework
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.9.0
dj-database-url==0.5.0
django-tailwind
django-allauth
requests
PyJWT==2.8.0
cryptography
python-decouple
mysqlclient
django-allauth
fastapi
firebase-admin
whitenoise
uvicorn
